File and data manage feature

+++ ProductTree/Processor.java
@@ -1,6 +1,6 @@
 abstract class Processor extends Product {
-    protected int coreNumber;
-    protected double clockRate; // Unit: GHz
-    protected int fabrication; // Unit: nm
-    protected int tdp; // Unit: W
+    protected Long coreNumber;
+    protected Double clockRate; // Unit: GHz
+    protected Long fabrication; // Unit: nm
+    protected Long tdp; // Unit: W
 }
